# TIFF read performance improvements