- The OME-TIFF reader was supposed to be caching open TIFF reader instances, but in practice it was not doing the caching properly, and was repeatedly re-opening the same TIFF file instead of using the cached one; it now opens each TIFF once only and correctly retrieves it from the cache
- The TIFF libtiff wrapper class was not caching IFD offsets, leading to repeated linear scans through the IFDs when accessing by index; now we cache all IFDs up front which means a single linear scan on open and then constant-time access

The first was the most painful, giving a 6× improvement with BBBC/mitocheck data (120s → 20s).  The second was less impressive at 1.1× (20s → 18s) but still useful.

The TIFF caching does need to also think about closing TIFFs (we currently leave them open indefinitely).  We probably want to use some sort of LRU ordered list and only keep a certain number open to avoid running out of file descriptors in pathological scenarios.  May be useful to wait on having user-settable options in the API to fine-tune this.

The offset caching could also be further optimised if we wished, for example by only caching up to the requested index on demand rather than up front which would make a slight improvement for usage involving access to a specific plane or range of planes only.

Testing: Check builds are green.  Run `ome-files info` on the BBBC or mitocheck files and note the time taken to complete.  This should result in a significant reduction when compared with previous builds.