@sbesson 2017-02-03
Initial review of this PR looks very positive from a functional behavior, the initialization time on Windows using the VC14 Release build against the `BBBC` OME-TIFF is reduced from 7min30 to 10s ! While reviewing, I realized we have not re-actived the latest OME Files builds so there is no direct comparison of the effect of this PR at the moment. I will perform the testing once more with [the latest OME Files artifacts](https://trello.com/c/lLiwhFIx/63-restore-latest-superbuilds) using both this file and our OME-TIFF sample files.