@sbesson 2017-02-06
Tested on CentOS 7.2 by opening the converted BBBC plate (6400 TiffData elements). The latest build initializes the reader in 1min47s while the merge build only takes 6s for the same operation.

This was confirmed on Windows/Ubuntu 16.04 by the pixel reading benchmarks for 0.2.3

![pixelsreading-0 2 0](https://cloud.githubusercontent.com/assets/1355463/22645654/fd2a9e8c-ec5f-11e6-8426-46fbfa80a916.png)

vs 0.3.0

![pixelsreading-0 3 0](https://cloud.githubusercontent.com/assets/1355463/22645668/0f8b3adc-ec60-11e6-86e3-43db89900445.png)

As discussed on Friday, the only impact of this change that might require some communication is the change of the protected map field in `lib/ome/files/in/OMETIFFReader.h`. As we are targetting a 0.3.0 release of OME Files, this is completely acceptable. Merging to move onto the next benchmarking issues.